chore: Migrate from structop to clap v3
`structopt` is now in maintenance mode, and clap v3 has integrated the
struct derive bits from the deprecated library. This PR migrates from
`structopt` to `clap` and flattens the subcommands so they don't require
the `--cmd` flag anymore.
